backend/requirements.txt
blinker==1.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
python-dotenv==0.21.1   # Updated version
requests==2.31.0
urllib3==2.0.7
Werkzeug==2.2.2
gunicorn==22.0.0
